requirements.txt
numpy
pandas
python-dotenv
setuptools
pymongo
certifi
pymongo[srv]==4.6.1
scikit-learn
dill
pyaml 
mlflow
dagshub
fastapi
uvicorn
huggingface_hub
joblib
-e .